tests/tokenswap: move async tests to new file
