+++ test/tokenswap/TokenSwap-test.ts
@@ -23,98 +23,43 @@ import {
     expectFulfilled,
     expectProperty,
     expectLength,
-    expectUndefined,
+    expectUndefined, expectBoolean, expectNull,
 } from "../helpers";
 
 
-describe("TokenSwap tests", function(this: Mocha.Suite) {
-    describe("Swap Rate tests", function(this: Mocha.Suite) {
+describe("TokenSwap -- Synchronous Tests", function(this: Mocha.Suite) {
+    describe("intermediateTokens tests", function(this: Mocha.Suite) {
         interface TestCase {
-            chainId:   number,
-            tokenFrom: Token,
-            tokenTo:   Token,
-            amountIn:  BigNumber,
-            wantError: boolean,
+            chainId: number;
+            token:   Token;
+            wantA:   Token;
+            wantB:   Token;
         }
 
-        const makeTestCase = (c: number, t1: Token, t2: Token, amt?: string, wantError?: boolean): TestCase =>
-            ({
-                chainId:   c,
-                tokenFrom: t1,
-                tokenTo:   t2,
-                amountIn:  getTestAmount(t1, c, amt),
-                wantError: wantError ?? false,
-            })
-
         const testCases: TestCase[] = [
-            makeTestCase(ChainId.ETH,        Tokens.DAI,        Tokens.USDC),
-            makeTestCase(ChainId.ETH,        Tokens.ETH,        Tokens.NETH, null, true),
-            makeTestCase(ChainId.OPTIMISM,   Tokens.WETH,       Tokens.NETH),
-            makeTestCase(ChainId.AVALANCHE,  Tokens.MIM,        Tokens.USDT, null, true),
-            makeTestCase(ChainId.BSC,        Tokens.BUSD,       Tokens.USDT),
-            makeTestCase(ChainId.BSC,        Tokens.NUSD,       Tokens.BUSD),
-            makeTestCase(ChainId.BSC,        Tokens.NUSD,       Tokens.DAI,  undefined,true),
-        ]
+            {chainId: ChainId.ETH,       token: Tokens.FRAX,     wantA: undefined,   wantB: Tokens.FRAX},
+            {chainId: ChainId.ETH,       token: Tokens.SYN_FRAX, wantA: undefined,   wantB: Tokens.FRAX},
+            {chainId: ChainId.HARMONY,   token: Tokens.FRAX,     wantA: undefined,   wantB: Tokens.SYN_FRAX},
+            {chainId: ChainId.ETH,       token: Tokens.NETH,     wantA: Tokens.NETH, wantB: Tokens.WETH},
+            {chainId: ChainId.ARBITRUM,  token: Tokens.NETH,     wantA: Tokens.NETH, wantB: Tokens.NETH},
+            {chainId: ChainId.BSC,       token: Tokens.BUSD,     wantA: Tokens.NUSD, wantB: Tokens.NUSD},
+            {chainId: ChainId.AVALANCHE, token: Tokens.GOHM,     wantA: Tokens.GOHM, wantB: Tokens.GOHM},
+        ];
 
         for (const tc of testCases) {
             const
-                titleSuffix: string = tc.wantError ? "should fail" : "should pass",
-                tokFrom: string     = tc.tokenFrom.symbol,
-                tokTo: string       = tc.tokenTo.symbol,
-                testTitle: string   = `for ${tokFrom} => ${tokTo} on ${Networks.networkName(tc.chainId)} ${titleSuffix}`,
-                testTitle1: string  = `calculateSwapRate ${testTitle}`,
-                testTitle2: string  = `buildSwapTokensTransaction ${testTitle}`,
-                testTitle3: string  = `swapSetup ${testTitle}`;
+                testPrefix: string = `intermediateTokens() with token ${tc.token.symbol} and Chain ID ${tc.chainId} should return`,
+                testWant:   string = `intermediateToken === ${tc.wantA?.symbol ?? 'undefined'}, bridgeConfigIntermediateToken === ${tc.wantB.symbol}`,
+                testTitle:  string = `${testPrefix} ${testWant}`;
 
-            let amountOut: BigNumber;
+            it(testTitle, function(this: Mocha.Context) {
+                const got = TokenSwap.intermediateTokens(tc.chainId, tc.token);
 
-            step(testTitle1, async function(this: Mocha.Context) {
-                this.timeout(DEFAULT_TEST_TIMEOUT);
+                typeof tc.wantA === "undefined"
+                    ? expectUndefined(tc.wantA, true)
+                    : expectBoolean(tc.wantA.isEqual(got.intermediateToken), true);
 
-                let prom: Promise<TokenSwap.EstimatedSwapRate> = Promise.resolve(TokenSwap.calculateSwapRate({
-                    chainId:   tc.chainId,
-                    tokenFrom: tc.tokenFrom,
-                    tokenTo:   tc.tokenTo,
-                    amountIn:  tc.amountIn,
-                })).then((res) => {
-                    amountOut = res.amountOut;
-                    return res
-                });
-
-                return tc.wantError
-                    ? await expectRejected(prom)
-                    : expectProperty(await prom, "amountOut").that.is.gt(Zero.toNumber())
-            })
-
-            step(testTitle2, async function(this: Mocha.Context) {
-                this.timeout(DEFAULT_TEST_TIMEOUT);
-
-                const args: TokenSwap.SwapTokensParams = {
-                    ...tc,
-                    minAmountOut: amountOut,
-                };
-
-                let prom = TokenSwap.buildSwapTokensTransaction(args);
-
-                return (await (
-                    tc.wantError
-                        ? expectRejected(prom)
-                        : expectFulfilled(prom)
-                ))
-            })
-
-            step(testTitle3, async function(this: Mocha.Context) {
-                this.timeout(DEFAULT_TEST_TIMEOUT);
-
-                let prom = TokenSwap.swapSetup(
-                    tc.tokenFrom, 
-                    tc.tokenTo, 
-                    tc.chainId,
-                )
-
-                return tc.wantError
-                    ? await expectRejected(prom)
-                    : expectProperty(await prom, "swapInstance").that.is.instanceof(Contract)
+                expectBoolean(tc.wantB.isEqual(got.bridgeConfigIntermediateToken), true);
             })
         }
     })
